Pass DUA vars through context instead of props
