+++ context/app/static/js/components/Detail/FileBrowserNode/FileBrowserNode.jsx
@@ -5,20 +5,14 @@ import FileBrowserFile from '../FileBrowserFile';
 import { Column } from './style';
 
 function FileBrowserNode(props) {
-  const { fileSubTree, depth, hasAgreedToDUA, openDUA } = props;
+  const { fileSubTree, depth } = props;
   return Object.entries(fileSubTree).map(([k, v]) => {
     // if the object contains array of files, display all files
     if (k === 'files') {
       return (
         <Column key={`${k}-${depth}`}>
           {v.map((file) => (
-            <FileBrowserFile
-              key={file.rel_path}
-              fileObj={file}
-              depth={depth}
-              hasAgreedToDUA={hasAgreedToDUA}
-              openDUA={openDUA}
-            />
+            <FileBrowserFile key={file.rel_path} fileObj={file} depth={depth} />
           ))}
         </Column>
       );
@@ -26,7 +20,7 @@ function FileBrowserNode(props) {
     // if the object contains additional directories, display dir and continue down
     return (
       <FileBrowserDirectory key={`${k}-${depth}`} dirName={k.slice(0, -1)} depth={depth}>
-        <FileBrowserNode fileSubTree={v} depth={depth + 1} hasAgreedToDUA={hasAgreedToDUA} openDUA={openDUA} />
+        <FileBrowserNode fileSubTree={v} depth={depth + 1} />
       </FileBrowserDirectory>
     );
   });
