+++ context/app/static/js/components/Detail/Files/context.js
@@ -1,0 +1,6 @@
+import React from 'react';
+
+// Stores the endpoints, `uuid`, and `display_doi` values.
+const FilesContext = React.createContext({});
+
+export default FilesContext;
